// agentark-web/src/shared/api/generated/control/models/cursorPageKnowledgeBase.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * AgentArk Public Control API
 * OpenAPI spec version: 1.0.0
 */
import type { CursorPageKnowledgeBaseItemsItem } from "./cursorPageKnowledgeBaseItemsItem";

export interface CursorPageKnowledgeBase {
  items: CursorPageKnowledgeBaseItemsItem[];
  /** @nullable */
  nextCursor: string | null;
  hasMore: boolean;
}