// agentark-web/src/shared/api/generated/control/models/cursorPageKnowledgeRevision.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * AgentArk Public Control API
 * OpenAPI spec version: 1.0.0
 */
import type { CursorPageKnowledgeRevisionItemsItem } from "./cursorPageKnowledgeRevisionItemsItem";

export interface CursorPageKnowledgeRevision {
  items: CursorPageKnowledgeRevisionItemsItem[];
  /** @nullable */
  nextCursor: string | null;
  hasMore: boolean;
}